Move TransactionVerifierConfig to its own submodule
