+++ chainstate/test-framework/src/tx_verification_strategy/randomized_strategy.rs
@@ -30,9 +30,9 @@ use common::{
 use crypto::random::{Rng, RngCore};
 use test_utils::random::{make_seedable_rng, Seed};
 use tx_verifier::transaction_verifier::{
-    error::ConnectTransactionError, flush::flush_to_storage,
+    config::TransactionVerifierConfig, error::ConnectTransactionError, flush::flush_to_storage,
     storage::TransactionVerifierStorageRef, BlockTransactableRef, BlockTransactableWithIndexRef,
-    Fee, Subsidy, TransactionVerifier, TransactionVerifierConfig, TransactionVerifierDelta,
+    Fee, Subsidy, TransactionVerifier, TransactionVerifierDelta,
 };
 use utils::tap_error_log::LogError;
 use utxo::UtxosView;
